--- client/src/app/components/Footer/index.js
@@ -1,26 +1,0 @@
-/* FOOTER
-*
-*
-*/
-
-import React from 'react'
-import styled from 'react-emotion'
-
-const Wrapper = styled("div")`
-  position: fixed;
-  display: flex;
-  justify-content: center;
-`
-
-const Social = styled("div")`
-  display: flex;
-  justify-content: space-between;
-`
-
-const Footer = () => {
-  <Wrapper>
-    <Social>
-      
-    </Social>
-  </Wrapper>
-}
