--- client/src/app/components/Form/index.js
@@ -1,12 +1,0 @@
-/*  FORM
-*   
-*
-*
-*/
-
-import React from 'react'
-import styled from 'react-emotion'
-
-const Footer = () => {
-  
-}
