--- client/src/app/components/Social/index.js
@@ -1,3 +1,0 @@
-import React from 'react'
-import styled from 'react-emotion'
-
