+++ client/src/components/Form/index.js
@@ -1,0 +1,10 @@
+/*  FORM
+*   
+*
+*
+*/
+
+import React from 'react';
+import styled from 'react-emotion';
+
+const Footer = () => {};
