+++ client/src/components/Social/index.js
@@ -1,0 +1,2 @@
+import React from 'react';
+import styled from 'react-emotion';
