+++ client/src/components/Vacancy/index.js
